refactor: extract toModuleSpecifier and accept URL inputs
Move the helper out of pathResolver so utils stay free of node builtins, and handle string | URL explicitly instead of String() casts.
